Update NutritionDataModel.swift
modified the data type of exerciseCalories (Int -> double)

+++ NutritionDataModel.swift
@@ -10,7 +10,7 @@ import Combine
 class NutritionDataModel: ObservableObject {
     @Published var calorieGoal: Int = 2700
     @Published var calorieIntake: Int = 0
-    @Published var exerciseCalories: Int = 500
+    @Published var exerciseCalories: Double = 500
     @Published var fatPercentage: Double = 0
     @Published var proteinPercentage: Double = 0
     @Published var carbPercentage: Double = 0
